Agustin develop

+++ finanzas/src/main/java/com/nocountry/finanzas/services/impl/UserServiceImpl.java
@@ -112,8 +112,13 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     @Override
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public void deleteUser(Long id) throws NotFoundException {
+
+        if(userRepository.findById(id).isPresent()){
+            userRepository.deleteById(id);
+        }else {
+            throw new NotFoundException("No se encuentra para eliminar el usuario con el id "+id);
+        }
     }
 
     @Transactional
@@ -122,7 +127,7 @@ public class UserServiceImpl implements UserService {
         Optional<User> userOptional = userRepository.findById(userLoggingDTO.getIdUser());
 
         if (userOptional.isEmpty()){
-            throw new NotFoundException("Could not found user");
+            throw new NotFoundException("No se pudo encontrar el usuario");
         }
 
         boolean isEmailCorrect = userOptional.get().getEmail().equalsIgnoreCase(userLoggingDTO.getEmail());
@@ -150,8 +155,7 @@ public class UserServiceImpl implements UserService {
         if (toSaving.getToSaving() > user.getTotalIncome()) {
             throw new NotFoundException("No hay fondos suficientes para invertir en ahorros.");
         }
-        user.setAccumulatedSavings(user.getAccumulatedSavings() + toSaving.getToSaving());
-        user.setTotalIncome(user.getTotalIncome() - toSaving.getToSaving());
+        user.setAccumulatedSavings(toSaving.getToSaving());
 
         userRepository.save(user);
         return Mapper.userToUserResponseDto(user);
@@ -161,7 +165,6 @@ public class UserServiceImpl implements UserService {
     public UserResponseDTO revertSavings(Long id) {
         User user = userRepository.findById(id).get();
 
-        user.setTotalIncome(user.getTotalIncome() + user.getAccumulatedSavings());
         user.setAccumulatedSavings(0.0);
 
         userRepository.save(user);
@@ -171,7 +174,7 @@ public class UserServiceImpl implements UserService {
     private void doEmailValidation(String email) throws InvalidEmailType {
 
         if (!emailValidator.isEmailValid(email)) {
-            throw new InvalidEmailType("El email ingresa no posee una estructura valida.");
+            throw new InvalidEmailType("El email ingresado no posee una estructura valida.");
         }
     }
